refactor: common module package cleanup
