+++ common/src/test/java/com/oing/domain/exception/ErrorCodeTest.java
@@ -1,5 +1,6 @@
 package com.oing.domain.exception;
 
+import com.oing.exception.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
