+++ family/src/main/java/com/oing/domain/model/FamilyInviteLink.java
@@ -1,5 +1,6 @@
 package com.oing.domain.model;
 
+import com.oing.domain.BaseEntity;
 import com.oing.domain.SerializableDeepLink;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
