+++ family/src/main/java/com/oing/exception/AlreadyInFamilyException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class AlreadyInFamilyException extends DomainException {
     public AlreadyInFamilyException() {
         super(ErrorCode.ALREADY_IN_FAMILY);
