+++ gateway/src/main/java/com/oing/config/SpringWebExceptionHandler.java
@@ -1,8 +1,8 @@
 package com.oing.config;
 
 import com.oing.domain.ErrorReportDTO;
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
+import com.oing.exception.DomainException;
+import com.oing.exception.ErrorCode;
 import com.oing.dto.response.ErrorResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
