+++ gateway/src/main/java/com/oing/config/filter/JwtAccessDeniedHandler.java
@@ -1,7 +1,7 @@
 package com.oing.config.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.oing.domain.exception.ErrorCode;
+import com.oing.exception.ErrorCode;
 import com.oing.dto.response.ErrorResponse;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
