+++ gateway/src/main/java/com/oing/domain/AppVersion.java
@@ -1,6 +1,5 @@
 package com.oing.domain;
 
-import com.oing.domain.model.BaseAuditEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
