+++ gateway/src/main/java/com/oing/domain/DeepLink.java
@@ -1,6 +1,5 @@
 package com.oing.domain;
 
-import com.oing.domain.model.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
