+++ member/src/main/java/com/oing/exception/MemberNotFoundException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class MemberNotFoundException extends DomainException {
 
     public MemberNotFoundException() {
