+++ member/src/main/java/com/oing/service/AuthService.java
@@ -7,8 +7,8 @@ import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.jwk.JWK;
 import com.oing.domain.SocialLoginProvider;
 import com.oing.domain.SocialLoginResult;
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
+import com.oing.exception.DomainException;
+import com.oing.exception.ErrorCode;
 import com.oing.dto.response.AppleKeyListResponse;
 import com.oing.dto.response.AppleKeyResponse;
 import com.oing.dto.response.KakaoAuthResponse;
