+++ post/src/main/java/com/oing/exception/DuplicatePostUploadException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class DuplicatePostUploadException extends DomainException {
     public DuplicatePostUploadException() {
         super(ErrorCode.DUPLICATE_POST_UPLOAD);
