+++ post/src/main/java/com/oing/exception/EmojiAlreadyExistsException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class EmojiAlreadyExistsException extends DomainException {
     public EmojiAlreadyExistsException() {
         super(ErrorCode.EMOJI_ALREADY_EXISTS);
