+++ post/src/main/java/com/oing/exception/EmojiNotFoundException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class EmojiNotFoundException extends DomainException {
     public EmojiNotFoundException() {
         super(ErrorCode.EMOJI_NOT_FOUND);
