+++ post/src/main/java/com/oing/exception/InvalidUploadTimeException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class InvalidUploadTimeException extends DomainException {
 
     public InvalidUploadTimeException() {
