+++ post/src/main/java/com/oing/exception/PostNotFoundException.java
@@ -1,8 +1,5 @@
 package com.oing.exception;
 
-import com.oing.domain.exception.DomainException;
-import com.oing.domain.exception.ErrorCode;
-
 public class PostNotFoundException extends DomainException {
     public PostNotFoundException() {
         super(ErrorCode.POST_NOT_FOUND);
